Break up test_random.py
