+++ pymc/tests/distributions/test_discrete.py
@@ -12,6 +12,7 @@
 #   See the License for the specific language governing permissions and
 #   limitations under the License.
 
+import functools as ft
 import sys
 import warnings
 
@@ -24,13 +25,15 @@ import scipy.stats as st
 
 from aeppl.logprob import ParameterValueError
 from aesara.compile.mode import Mode
-from aesara.tensor.var import TensorVariable
+from aesara.tensor import TensorVariable
 
 import pymc as pm
 
 from pymc.aesaraf import floatX
 from pymc.distributions import joint_logp, logcdf, logp
+from pymc.distributions.discrete import _OrderedLogistic, _OrderedProbit
 from pymc.tests.distributions.util import (
+    BaseTestDistributionRandom,
     Bool,
     Domain,
     I,
@@ -50,6 +53,8 @@ from pymc.tests.distributions.util import (
     check_logcdf,
     check_logp,
     check_selfconsistency_discrete_logcdf,
+    seeded_numpy_distribution_builder,
+    seeded_scipy_distribution_builder,
 )
 from pymc.vartypes import discrete_types
 
@@ -572,18 +577,6 @@ def test_orderedlogistic_dimensions(shape):
     assert np.allclose(ologp, expected)
 
 
-def test_ordered_multinomial_probs():
-    with pm.Model() as m:
-        pm.OrderedMultinomial("om_p", n=1000, cutpoints=np.array([-2, 0, 2]), eta=0)
-        pm.OrderedMultinomial(
-            "om_no_p", n=1000, cutpoints=np.array([-2, 0, 2]), eta=0, compute_p=False
-        )
-    assert len(m.deterministics) == 1
-
-    x = pm.OrderedMultinomial.dist(n=1000, cutpoints=np.array([-2, 0, 2]), eta=0)
-    assert isinstance(x, TensorVariable)
-
-
 def test_ordered_logistic_probs():
     with pm.Model() as m:
         pm.OrderedLogistic("ol_p", cutpoints=np.array([-2, 0, 2]), eta=0)
@@ -863,3 +856,280 @@ class TestMoments:
         with pm.Model() as model:
             pm.ZeroInflatedNegativeBinomial("x", psi=psi, mu=mu, alpha=alpha, size=size)
         assert_moment_is_expected(model, expected)
+
+
+class TestDiscreteWeibull(BaseTestDistributionRandom):
+    def discrete_weibul_rng_fn(self, size, q, beta, uniform_rng_fct):
+        return np.ceil(np.power(np.log(1 - uniform_rng_fct(size=size)) / np.log(q), 1.0 / beta)) - 1
+
+    def seeded_discrete_weibul_rng_fn(self):
+        uniform_rng_fct = ft.partial(
+            getattr(np.random.RandomState, "uniform"), self.get_random_state()
+        )
+        return ft.partial(self.discrete_weibul_rng_fn, uniform_rng_fct=uniform_rng_fct)
+
+    pymc_dist = pm.DiscreteWeibull
+    pymc_dist_params = {"q": 0.25, "beta": 2.0}
+    expected_rv_op_params = {"q": 0.25, "beta": 2.0}
+    reference_dist_params = {"q": 0.25, "beta": 2.0}
+    reference_dist = seeded_discrete_weibul_rng_fn
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_pymc_draws_match_reference",
+        "check_rv_size",
+    ]
+
+
+class TestBinomial(BaseTestDistributionRandom):
+    pymc_dist = pm.Binomial
+    pymc_dist_params = {"n": 100, "p": 0.33}
+    expected_rv_op_params = {"n": 100, "p": 0.33}
+    checks_to_run = ["check_pymc_params_match_rv_op"]
+
+
+class TestLogitBinomial(BaseTestDistributionRandom):
+    pymc_dist = pm.Binomial
+    pymc_dist_params = {"n": 100, "logit_p": 0.5}
+    expected_rv_op_params = {"n": 100, "p": sp.expit(0.5)}
+    tests_to_run = ["check_pymc_params_match_rv_op"]
+
+    @pytest.mark.parametrize(
+        "n, p, logit_p, expected",
+        [
+            (5, None, None, "Must specify either p or logit_p."),
+            (5, 0.5, 0.5, "Can't specify both p and logit_p."),
+        ],
+    )
+    def test_binomial_init_fail(self, n, p, logit_p, expected):
+        with pm.Model() as model:
+            with pytest.raises(ValueError, match=f"Incompatible parametrization. {expected}"):
+                pm.Binomial("x", n=n, p=p, logit_p=logit_p)
+
+
+class TestNegativeBinomial(BaseTestDistributionRandom):
+    pymc_dist = pm.NegativeBinomial
+    pymc_dist_params = {"n": 100, "p": 0.33}
+    expected_rv_op_params = {"n": 100, "p": 0.33}
+    checks_to_run = ["check_pymc_params_match_rv_op"]
+
+
+class TestNegativeBinomialMuSigma(BaseTestDistributionRandom):
+    pymc_dist = pm.NegativeBinomial
+    pymc_dist_params = {"mu": 5.0, "alpha": 8.0}
+    expected_n, expected_p = pm.NegativeBinomial.get_n_p(
+        mu=pymc_dist_params["mu"],
+        alpha=pymc_dist_params["alpha"],
+        n=None,
+        p=None,
+    )
+    expected_rv_op_params = {"n": expected_n, "p": expected_p}
+    checks_to_run = ["check_pymc_params_match_rv_op"]
+
+
+class TestBernoulli(BaseTestDistributionRandom):
+    pymc_dist = pm.Bernoulli
+    pymc_dist_params = {"p": 0.33}
+    expected_rv_op_params = {"p": 0.33}
+    reference_dist_params = {"p": 0.33}
+    reference_dist = seeded_scipy_distribution_builder("bernoulli")
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_pymc_draws_match_reference",
+    ]
+
+
+class TestBernoulliLogitP(BaseTestDistributionRandom):
+    pymc_dist = pm.Bernoulli
+    pymc_dist_params = {"logit_p": 1.0}
+    expected_rv_op_params = {"p": sp.expit(1.0)}
+    checks_to_run = ["check_pymc_params_match_rv_op"]
+
+
+class TestPoisson(BaseTestDistributionRandom):
+    pymc_dist = pm.Poisson
+    pymc_dist_params = {"mu": 4.0}
+    expected_rv_op_params = {"mu": 4.0}
+    checks_to_run = ["check_pymc_params_match_rv_op"]
+
+
+class TestCategorical(BaseTestDistributionRandom):
+    pymc_dist = pm.Categorical
+    pymc_dist_params = {"p": np.array([0.28, 0.62, 0.10])}
+    expected_rv_op_params = {"p": np.array([0.28, 0.62, 0.10])}
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_rv_size",
+    ]
+
+
+class TestLogitCategorical(BaseTestDistributionRandom):
+    pymc_dist = pm.Categorical
+    pymc_dist_params = {"logit_p": np.array([[0.28, 0.62, 0.10], [0.28, 0.62, 0.10]])}
+    expected_rv_op_params = {
+        "p": sp.softmax(np.array([[0.28, 0.62, 0.10], [0.28, 0.62, 0.10]]), axis=-1)
+    }
+    tests_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_rv_size",
+    ]
+
+    @pytest.mark.parametrize(
+        "p, logit_p, expected",
+        [
+            (None, None, "Must specify either p or logit_p."),
+            (0.5, 0.5, "Can't specify both p and logit_p."),
+        ],
+    )
+    def test_categorical_init_fail(self, p, logit_p, expected):
+        with pm.Model() as model:
+            with pytest.raises(ValueError, match=f"Incompatible parametrization. {expected}"):
+                pm.Categorical("x", p=p, logit_p=logit_p)
+
+
+class TestGeometric(BaseTestDistributionRandom):
+    pymc_dist = pm.Geometric
+    pymc_dist_params = {"p": 0.9}
+    expected_rv_op_params = {"p": 0.9}
+    checks_to_run = ["check_pymc_params_match_rv_op"]
+
+
+class TestHyperGeometric(BaseTestDistributionRandom):
+    pymc_dist = pm.HyperGeometric
+    pymc_dist_params = {"N": 20, "k": 12, "n": 5}
+    expected_rv_op_params = {
+        "ngood": pymc_dist_params["k"],
+        "nbad": pymc_dist_params["N"] - pymc_dist_params["k"],
+        "nsample": pymc_dist_params["n"],
+    }
+    reference_dist_params = expected_rv_op_params
+    reference_dist = seeded_numpy_distribution_builder("hypergeometric")
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_pymc_draws_match_reference",
+    ]
+
+
+class TestBetaBinomial(BaseTestDistributionRandom):
+    pymc_dist = pm.BetaBinomial
+    pymc_dist_params = {"alpha": 2.0, "beta": 1.0, "n": 5}
+    expected_rv_op_params = {"n": 5, "alpha": 2.0, "beta": 1.0}
+    reference_dist_params = {"n": 5, "a": 2.0, "b": 1.0}
+    reference_dist = seeded_scipy_distribution_builder("betabinom")
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_pymc_draws_match_reference",
+        "check_rv_size",
+    ]
+
+
+class TestDiscreteUniform(BaseTestDistributionRandom):
+    def discrete_uniform_rng_fn(self, size, lower, upper, rng):
+        return st.randint.rvs(lower, upper + 1, size=size, random_state=rng)
+
+    pymc_dist = pm.DiscreteUniform
+    pymc_dist_params = {"lower": -1, "upper": 9}
+    expected_rv_op_params = {"lower": -1, "upper": 9}
+    reference_dist_params = {"lower": -1, "upper": 9}
+    reference_dist = lambda self: ft.partial(
+        self.discrete_uniform_rng_fn, rng=self.get_random_state()
+    )
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_pymc_draws_match_reference",
+        "check_rv_size",
+    ]
+
+
+class TestDiracDelta(BaseTestDistributionRandom):
+    def diracdelta_rng_fn(self, size, c):
+        if size is None:
+            return c
+        return np.full(size, c)
+
+    pymc_dist = pm.DiracDelta
+    pymc_dist_params = {"c": 3}
+    expected_rv_op_params = {"c": 3}
+    reference_dist_params = {"c": 3}
+    reference_dist = lambda self: self.diracdelta_rng_fn
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_pymc_draws_match_reference",
+        "check_rv_size",
+    ]
+
+    @pytest.mark.parametrize("floatX", ["float32", "float64"])
+    @pytest.mark.xfail(
+        sys.platform == "win32", reason="https://github.com/aesara-devs/aesara/issues/871"
+    )
+    def test_dtype(self, floatX):
+        with aesara.config.change_flags(floatX=floatX):
+            assert pm.DiracDelta.dist(2**4).dtype == "int8"
+            assert pm.DiracDelta.dist(2**16).dtype == "int32"
+            assert pm.DiracDelta.dist(2**32).dtype == "int64"
+            assert pm.DiracDelta.dist(2.0).dtype == floatX
+
+
+class TestOrderedLogistic(BaseTestDistributionRandom):
+    pymc_dist = _OrderedLogistic
+    pymc_dist_params = {"eta": 0, "cutpoints": np.array([-2, 0, 2])}
+    expected_rv_op_params = {"p": np.array([0.11920292, 0.38079708, 0.38079708, 0.11920292])}
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_rv_size",
+    ]
+
+    @pytest.mark.parametrize(
+        "eta, cutpoints, expected",
+        [
+            (0, [-2.0, 0, 2.0], (4,)),
+            ([-1], [-2.0, 0, 2.0], (1, 4)),
+            ([1.0, -2.0], [-1.0, 0, 1.0], (2, 4)),
+            (np.zeros((3, 2)), [-2.0, 0, 1.0], (3, 2, 4)),
+            (np.ones((5, 2)), [[-2.0, 0, 1.0], [-1.0, 0, 1.0]], (5, 2, 4)),
+            (np.ones((3, 5, 2)), [[-2.0, 0, 1.0], [-1.0, 0, 1.0]], (3, 5, 2, 4)),
+        ],
+    )
+    def test_shape_inputs(self, eta, cutpoints, expected):
+        """
+        This test checks when providing different shapes for `eta` parameters.
+        """
+        categorical = _OrderedLogistic.dist(
+            eta=eta,
+            cutpoints=cutpoints,
+        )
+        p = categorical.owner.inputs[3].eval()
+        assert p.shape == expected
+
+
+class TestOrderedProbit(BaseTestDistributionRandom):
+    pymc_dist = _OrderedProbit
+    pymc_dist_params = {"eta": 0, "cutpoints": np.array([-2, 0, 2])}
+    expected_rv_op_params = {"p": np.array([0.02275013, 0.47724987, 0.47724987, 0.02275013])}
+    checks_to_run = [
+        "check_pymc_params_match_rv_op",
+        "check_rv_size",
+    ]
+
+    @pytest.mark.parametrize(
+        "eta, cutpoints, sigma, expected",
+        [
+            (0, [-2.0, 0, 2.0], 1.0, (4,)),
+            ([-1], [-1.0, 0, 2.0], [2.0], (1, 4)),
+            ([1.0, -2.0], [-1.0, 0, 1.0], 1.0, (2, 4)),
+            ([1.0, -2.0, 3.0], [-1.0, 0, 2.0], np.ones((1, 3)), (1, 3, 4)),
+            (np.zeros((2, 3)), [-2.0, 0, 1.0], [1.0, 2.0, 5.0], (2, 3, 4)),
+            (np.ones((2, 3)), [-1.0, 0, 1.0], np.ones((2, 3)), (2, 3, 4)),
+            (np.zeros((5, 2)), [[-2, 0, 1], [-1, 0, 1]], np.ones((2, 5, 2)), (2, 5, 2, 4)),
+        ],
+    )
+    def test_shape_inputs(self, eta, cutpoints, sigma, expected):
+        """
+        This test checks when providing different shapes for `eta` and `sigma` parameters.
+        """
+        categorical = _OrderedProbit.dist(
+            eta=eta,
+            cutpoints=cutpoints,
+            sigma=sigma,
+        )
+        p = categorical.owner.inputs[3].eval()
+        assert p.shape == expected
