+++ pymc/tests/distributions/util.py
@@ -1,24 +1,28 @@
+import functools as ft
 import itertools as it
 
 from contextlib import ExitStack as does_not_raise
+from typing import Callable, List, Optional
 
 import aesara
 import aesara.tensor as at
 import numpy as np
 import numpy.random as nr
+import numpy.testing as npt
 import pytest
 import scipy.special as sp
+import scipy.stats as st
 
 from aeppl.logprob import ParameterValueError
 from aesara.compile.mode import Mode
 
 import pymc as pm
 
-from pymc.aesaraf import floatX
+from pymc.aesaraf import change_rv_size, compile_pymc, floatX
 from pymc.distributions import logcdf, logp
 from pymc.distributions.logprob import joint_logp
 from pymc.initial_point import make_initial_point_fn
-from pymc.tests.helpers import select_by_precision
+from pymc.tests.helpers import SeededTest, select_by_precision
 
 
 def product(domains, n_samples=-1):
@@ -312,7 +316,7 @@ def check_logp(
         pt_d = _model_input_dict(model, param_vars, pt)
         pt_logp = pm.Point(pt_d, model=model)
         pt_ref = pm.Point(pt, filter_model_vars=False, model=model)
-        np.testing.assert_almost_equal(
+        npt.assert_almost_equal(
             logp_pymc(pt_logp),
             logp_reference(pt_ref),
             decimal=decimal,
@@ -366,7 +370,7 @@ def check_logp(
         if invalid_value is None:
             continue
         with aesara.config.change_flags(mode=Mode("py")):
-            np.testing.assert_equal(
+            npt.assert_equal(
                 logp(valid_dist, invalid_value).eval(),
                 -np.inf,
                 err_msg=str(invalid_value),
@@ -452,7 +456,7 @@ def check_logcdf(
             pymc_eval = pymc_logcdf({"value": value})
 
             params["value"] = value  # for displaying in err_msg
-            np.testing.assert_almost_equal(
+            npt.assert_almost_equal(
                 pymc_eval,
                 scipy_eval,
                 decimal=decimal,
@@ -493,7 +497,7 @@ def check_logcdf(
     if np.isfinite(domain.lower):
         below_domain = domain.lower - 1
         with aesara.config.change_flags(mode=Mode("py")):
-            np.testing.assert_equal(
+            npt.assert_equal(
                 logcdf(valid_dist, below_domain).eval(),
                 -np.inf,
                 err_msg=str(below_domain),
@@ -503,7 +507,7 @@ def check_logcdf(
     if np.isfinite(domain.upper):
         above_domain = domain.upper + 1
         with aesara.config.change_flags(mode=Mode("py")):
-            np.testing.assert_equal(
+            npt.assert_equal(
                 logcdf(valid_dist, above_domain).eval(),
                 0,
                 err_msg=str(above_domain),
@@ -551,7 +555,7 @@ def check_selfconsistency_discrete_logcdf(
             param_vars[param_name].set_value(param_value)
 
         with aesara.config.change_flags(mode=Mode("py")):
-            np.testing.assert_almost_equal(
+            npt.assert_almost_equal(
                 dist_logcdf({"value": value}),
                 sp.logsumexp([dist_logp({"value": value}) for value in values]),
                 decimal=decimal,
@@ -579,3 +583,264 @@ def assert_moment_is_expected(model, expected, check_finite_logp=True):
     if check_finite_logp:
         logp_moment = joint_logp(model["x"], at.constant(moment), transformed=False).eval()
         assert np.isfinite(logp_moment)
+
+
+def pymc_random(
+    dist,
+    paramdomains,
+    ref_rand,
+    valuedomain=None,
+    size=10000,
+    alpha=0.05,
+    fails=10,
+    extra_args=None,
+    model_args=None,
+    change_rv_size_fn=change_rv_size,
+):
+    if valuedomain is None:
+        valuedomain = Domain([0], edges=(None, None))
+
+    if model_args is None:
+        model_args = {}
+
+    model, param_vars = build_model(dist, valuedomain, paramdomains, extra_args)
+    model_dist = change_rv_size_fn(model.named_vars["value"], size, expand=True)
+    pymc_rand = compile_pymc([], model_dist)
+
+    domains = paramdomains.copy()
+    for pt in product(domains, n_samples=100):
+        pt = pm.Point(pt, model=model)
+        pt.update(model_args)
+
+        # Update the shared parameter variables in `param_vars`
+        for k, v in pt.items():
+            nv = param_vars.get(k, model.named_vars.get(k))
+            if nv.name in param_vars:
+                param_vars[nv.name].set_value(v)
+
+        p = alpha
+        # Allow KS test to fail (i.e., the samples be different)
+        # a certain number of times. Crude, but necessary.
+        f = fails
+        while p <= alpha and f > 0:
+            s0 = pymc_rand()
+            s1 = floatX(ref_rand(size=size, **pt))
+            _, p = st.ks_2samp(np.atleast_1d(s0).flatten(), np.atleast_1d(s1).flatten())
+            f -= 1
+        assert p > alpha, str(pt)
+
+
+def pymc_random_discrete(
+    dist,
+    paramdomains,
+    valuedomain=None,
+    ref_rand=None,
+    size=100000,
+    alpha=0.05,
+    fails=20,
+):
+    if valuedomain is None:
+        valuedomain = Domain([0], edges=(None, None))
+
+    model, param_vars = build_model(dist, valuedomain, paramdomains)
+    model_dist = change_rv_size(model.named_vars["value"], size, expand=True)
+    pymc_rand = compile_pymc([], model_dist)
+
+    domains = paramdomains.copy()
+    for pt in product(domains, n_samples=100):
+        pt = pm.Point(pt, model=model)
+        p = alpha
+
+        # Update the shared parameter variables in `param_vars`
+        for k, v in pt.items():
+            nv = param_vars.get(k, model.named_vars.get(k))
+            if nv.name in param_vars:
+                param_vars[nv.name].set_value(v)
+
+        # Allow Chisq test to fail (i.e., the samples be different)
+        # a certain number of times.
+        f = fails
+        while p <= alpha and f > 0:
+            o = pymc_rand()
+            e = intX(ref_rand(size=size, **pt))
+            o = np.atleast_1d(o).flatten()
+            e = np.atleast_1d(e).flatten()
+            bins = min(20, max(len(set(e)), len(set(o))))
+            range = (min(min(e), min(o)), max(max(e), max(o)))
+            observed, _ = np.histogram(o, bins=bins, range=range)
+            expected, _ = np.histogram(e, bins=bins, range=range)
+            if np.all(observed == expected):
+                p = 1.0
+            else:
+                _, p = st.chisquare(observed + 1, expected + 1)
+            f -= 1
+        assert p > alpha, str(pt)
+
+
+class BaseTestDistributionRandom(SeededTest):
+    """
+    This class provides a base for tests that new RandomVariables are correctly
+    implemented, and that the mapping of parameters between the PyMC
+    Distribution and the respective RandomVariable is correct.
+
+    Three default tests are provided which check:
+    1. Expected inputs are passed to the `rv_op` by the `dist` `classmethod`,
+    via `check_pymc_params_match_rv_op`
+    2. Expected (exact) draws are being returned, via
+    `check_pymc_draws_match_reference`
+    3. Shape variable inference is correct, via `check_rv_size`
+
+    Each desired test must be referenced by name in `tests_to_run`, when
+    subclassing this distribution. Custom tests can be added to each class as
+    well. See `TestFlat` for an example.
+
+    Additional tests should be added for each optional parametrization of the
+    distribution. In this case it's enough to include the test
+    `check_pymc_params_match_rv_op` since only this differs.
+
+    Note on `check_rv_size` test:
+        Custom input sizes (and expected output shapes) can be defined for the
+        `check_rv_size` test, by adding the optional class attributes
+        `sizes_to_check` and `sizes_expected`:
+
+        ```python
+        sizes_to_check = [None, (1), (2, 3)]
+        sizes_expected = [(3,), (1, 3), (2, 3, 3)]
+        tests_to_run = ["check_rv_size"]
+        ```
+
+        This is usually needed for Multivariate distributions. You can see an
+        example in `TestDirichlet`
+
+
+    Notes on `check_pymcs_draws_match_reference` test:
+
+        The `check_pymcs_draws_match_reference` is a very simple test for the
+        equality of draws from the `RandomVariable` and the exact same python
+        function, given the  same inputs and random seed. A small number
+        (`size=15`) is checked. This is not supposed to be a test for the
+        correctness of the random generator. The latter kind of test
+        (if warranted) can be performed with the aid of `pymc_random` and
+        `pymc_random_discrete` methods in this file, which will perform an
+        expensive statistical comparison between the RandomVariable `rng_fn`
+        and a reference Python function. This kind of test only makes sense if
+        there is a good independent generator reference (i.e., not just the same
+        composition of numpy / scipy python calls that is done inside `rng_fn`).
+
+        Finally, when your `rng_fn` is doing something more than just calling a
+        `numpy` or `scipy` method, you will need to setup an equivalent seeded
+        function with which to compare for the exact draws (instead of relying on
+        `seeded_[scipy|numpy]_distribution_builder`). You can find an example
+        in the `TestWeibull`, whose `rng_fn` returns
+        `beta * np.random.weibull(alpha, size=size)`.
+
+    """
+
+    pymc_dist: Optional[Callable] = None
+    pymc_dist_params: Optional[dict] = None
+    reference_dist: Optional[Callable] = None
+    reference_dist_params: Optional[dict] = None
+    expected_rv_op_params: Optional[dict] = None
+    checks_to_run = []
+    size = 15
+    decimal = select_by_precision(float64=6, float32=3)
+
+    sizes_to_check: Optional[List] = None
+    sizes_expected: Optional[List] = None
+    repeated_params_shape = 5
+
+    def test_distribution(self):
+        self.validate_tests_list()
+        if self.pymc_dist == pm.Wishart:
+            with pytest.warns(UserWarning, match="can currently not be used for MCMC sampling"):
+                self._instantiate_pymc_rv()
+        else:
+            self._instantiate_pymc_rv()
+        if self.reference_dist is not None:
+            self.reference_dist_draws = self.reference_dist()(
+                size=self.size, **self.reference_dist_params
+            )
+        for check_name in self.checks_to_run:
+            if check_name.startswith("test_"):
+                raise ValueError(
+                    "Custom check cannot start with `test_` or else it will be executed twice."
+                )
+            if self.pymc_dist == pm.Wishart and check_name.startswith("check_rv_size"):
+                with pytest.warns(UserWarning, match="can currently not be used for MCMC sampling"):
+                    getattr(self, check_name)()
+            else:
+                getattr(self, check_name)()
+
+    def _instantiate_pymc_rv(self, dist_params=None):
+        params = dist_params if dist_params else self.pymc_dist_params
+        self.pymc_rv = self.pymc_dist.dist(
+            **params, size=self.size, rng=aesara.shared(self.get_random_state(reset=True))
+        )
+
+    def check_pymc_draws_match_reference(self):
+        # need to re-instantiate it to make sure that the order of drawings match the reference distribution one
+        # self._instantiate_pymc_rv()
+        npt.assert_array_almost_equal(
+            self.pymc_rv.eval(), self.reference_dist_draws, decimal=self.decimal
+        )
+
+    def check_pymc_params_match_rv_op(self):
+        aesara_dist_inputs = self.pymc_rv.get_parents()[0].inputs[3:]
+        assert len(self.expected_rv_op_params) == len(aesara_dist_inputs)
+        for (expected_name, expected_value), actual_variable in zip(
+            self.expected_rv_op_params.items(), aesara_dist_inputs
+        ):
+
+            # Add additional line to evaluate symbolic inputs to distributions
+            if isinstance(expected_value, aesara.tensor.Variable):
+                expected_value = expected_value.eval()
+
+            npt.assert_almost_equal(expected_value, actual_variable.eval(), decimal=self.decimal)
+
+    def check_rv_size(self):
+        # test sizes
+        sizes_to_check = self.sizes_to_check or [None, (), 1, (1,), 5, (4, 5), (2, 4, 2)]
+        sizes_expected = self.sizes_expected or [(), (), (1,), (1,), (5,), (4, 5), (2, 4, 2)]
+        for size, expected in zip(sizes_to_check, sizes_expected):
+            pymc_rv = self.pymc_dist.dist(**self.pymc_dist_params, size=size)
+            expected_symbolic = tuple(pymc_rv.shape.eval())
+            actual = pymc_rv.eval().shape
+            assert actual == expected_symbolic
+            assert expected_symbolic == expected
+
+        # test multi-parameters sampling for univariate distributions (with univariate inputs)
+        if (
+            self.pymc_dist.rv_op.ndim_supp == 0
+            and self.pymc_dist.rv_op.ndims_params
+            and sum(self.pymc_dist.rv_op.ndims_params) == 0
+        ):
+            params = {
+                k: p * np.ones(self.repeated_params_shape) for k, p in self.pymc_dist_params.items()
+            }
+            self._instantiate_pymc_rv(params)
+            sizes_to_check = [None, self.repeated_params_shape, (5, self.repeated_params_shape)]
+            sizes_expected = [
+                (self.repeated_params_shape,),
+                (self.repeated_params_shape,),
+                (5, self.repeated_params_shape),
+            ]
+            for size, expected in zip(sizes_to_check, sizes_expected):
+                pymc_rv = self.pymc_dist.dist(**params, size=size)
+                expected_symbolic = tuple(pymc_rv.shape.eval())
+                actual = pymc_rv.eval().shape
+                assert actual == expected_symbolic == expected
+
+    def validate_tests_list(self):
+        assert len(self.checks_to_run) == len(
+            set(self.checks_to_run)
+        ), "There are duplicates in the list of tests_to_run"
+
+
+def seeded_scipy_distribution_builder(dist_name: str) -> Callable:
+    return lambda self: ft.partial(getattr(st, dist_name).rvs, random_state=self.get_random_state())
+
+
+def seeded_numpy_distribution_builder(dist_name: str) -> Callable:
+    return lambda self: ft.partial(
+        getattr(np.random.RandomState, dist_name), self.get_random_state()
+    )
